Add CK_ObjectTree::getInputTree and ::getModuleList
